requirements.txt
wheel==0.33.1
aiofiles==0.4.0
alembic==1.0.7
astroid==2.1.0
backcall==0.1.0
decorator==4.3.2
httptools==0.0.11
ipython==7.2.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.2
lazy-object-proxy==1.3.1
Mako==1.0.7
manage.py==0.2.10
MarkupSafe==1.1.0
mccabe==0.6.1
multidict==4.5.2
numpy==1.16.2
pandas==0.24.1
parso==0.3.3
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.8
psycopg2==2.7.7
psycopg2-binary==2.7.7
ptyprocess==0.6.0
Pygments==2.3.1
pylint==2.2.2
pymemcache==2.1.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2018.9
sanic==18.12.0
six==1.12.0
SQLAlchemy==1.2.17
traitlets==4.3.2
typed-ast==1.3.1
ujson==1.35
uvloop==0.12.0
wcwidth==0.1.7
websockets==6.0
wrapt==1.11.1
